web_requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
Pillow>=9.0.0
numpy>=1.21.0
requests>=2.28.0 